+++ changelog.md
@@ -20,6 +20,10 @@ While all _Major_ releases should be reviewed, our only _large_ releases are **v
 #### 14.11.2 [Jan 2024]
 - **[new]** -  support for frozen lex in plugin object #1080
 - **[fix]** - toggling options in .json()
+- **[new]** - .join() and .joinIf() methods
+- **[change]** - internal typescript version
+
+https://observablehq.com/@spencermountain/compromise-split
 -->
 
 #### 14.11.1 [Jan 2024]